Set up unit testing with pytest and Coverage.py
Add a test for the “hello world” function, to make sure everything
is set up properly.

For now, the application uses a custom test environment rather than
Hatch’s build-in testing setup. The latter is useful for testing
against a number of different Python versions prior to release, but
is somewhat inflexible, and during development, that inflexibility
is annoying, and it’s sufficient to test against a single Python
version.