Handle errors due to bad `getxattr()` arguments
This shouldn’t currently be possible, so there is no specific exception
class for this case, but the `CannotReadWhereFromValue` exception that
would be raised if that case should, after all, occur, would correctly
specify the error name as `EINVAL` now, at least, and the situation has
received documentation.

Also, check error codes and error names in all existing tests, use this
opportunity to test unexpected errors.