Handle `EISDIR`
It’s not clear under what conditions `getxattr()` would use this error
code. The name suggests the error would occur when trying to read an
extended attribute of a directory, but reading the “where from” value
of a directory works just fine (on an APFS volume, at least).

In any case, should the error occur, the application will now treat it
like `EPERM` and raise `UnsupportedFileSystemObject`.

Also, added a test that reads the “where from” value of a directory.