Handle errors due to unreadable files
Add an `FileNotReadable` exception class to handle `getxattr()` failing
with `EACCES`.

Also, move the test environment into the application’s main directory.

(Being in the `tests` directory meant that pytest, mypy, and Ruff were
searching the test environment for tests to import and code to check,
and that caused issues now that it contains directories that cannot be
read. The tools could, instead have been configured to ignore the
directory, but the test environment was out of place in the `tests`
directory in any case, and would be more so if the tests were organized
into subdirectories, which they might be at some point.)