Delete test environments with looping symlinks
Previously, an error would occur during the deletion of the test
environment if it contained symlinks that would cause an endless loop
if indiscriminately followed.

Now, symlinks named “loop.forever” are deleted without being followed,
and other symlinks are ignored.

This will become useful when testing the `ELOOP` error, which is next
on the list.