Make malformed “where from” values an error
If `read_where_from_value()` reads a “where from” value that isn’t
a nonempty list of strings, it raises `MalformedWhereFromValue`
now.

I haven’t been able to find any particularly clear documentation as to
what type the “where from” value is supposed to have, but according to
http://developer.apple.com/documentation/coreservices/kmditemwherefroms
it’s a list of strings, so I‘m reasonably comfortable with treating all
other types as an error.