Restructure the tests
Move the tests for `wherefrom.read` module into a separate package,
split the tests for unexpected values into its own module, and
reorganize the remaining tests for better readability.

Also, rename `MalformedWhereFromValue` to `UnexpectedWhereFromValue`
(malformed values are values that cannot even be parsed), and rename
`UnsupportedFileName` to `UnsupportedPath` (the issue may be the
length of the path rather than the file name).