Split parsing into its own module
Move the function for parsing the binary “where from” value out of
`wherefrom.read`, into `wherefrom.parse`.

That module is rather small and won’t become much larger, so it may
seem that it’s not really worthwhile to move the functionality out of
`read`, but this move positively affects the organization of the rest
of the code; test cases and exception classes related to parsing no
longer get lost among massive amounts of code related to reading the
value, which improves the organization of these files and removes the
need to split them into pieces that clash with the organization of the
code they’re supporting.

Accordingly, `tests.read.happy_tests` and `.unhappy_getxattr_tests`
have been merged and moved back into `tests.read`, and the awkward
`tests.read.unhappy_value_tests` module is now part of `tests.parse`.

Also, renamed `wherefrom.errors` into `wherefrom.exceptions` and moved
most of the exception classes out of it. The one exception related to
parsing the value is now in `wherefrom.parse`, and the large number of
exceptions related to reading the binary value are now in the new
`wherefrom.readexceptions` module, which has a somewhat unfortunate
name, but keeps the exceptions out of the way.