Rename a number of exceptions
Base classes for exception sub-hierarchies are now consistently named
`⟨Topic⟩Error`, and leaf exceptions are called either `No⟨GoodThing⟩`
(as in `NoReadPermission`) or `⟨BadThing⟩` (as in `TooManySymlinks`).

`NoSuchFile` would have fit into the `No⟨GoodThing⟩` pattern, but has
been renamed to `MissingFile` nonetheless, since the “such” would make
no sense if the exception were reported to the user or propagated to
parts of the application that have no direct involvement with the file
in question.