Refactor error handling
Exceptions are now defined in the `wherefrom.exceptions` package.

Some exception classes have been renamed so that they can be reused for
handling the equivalent issue during while walking the directory tree.

Also, there is now a (terribly over-designed) mechanism that allows
exception classes to be registered for error codes (and, optionally,
operations) and automatically creates a suitable exception for a given
error code, path, and operation name.