Ensure that the exception hierarchy is complete
The docstring of the `wherefrom.exceptions.base` module shows the
application’s exception hierarchy, but there is always the danger
that the developer forgets to add a newly-created exception class
to the exception hierarchy.

To address that issue, there is now a test that verifies that the
names of all `WhereFromException` subclasses are mentioned in the
hierarchy.