Mark exception classes as ignorable
Add a class attribute to `FileError` that can be overridden by its
subclasses to indicate that instances of that subclass can be ignored
while walking the file system.

This moves another bit of exception information into the `exceptions`
package, where it belongs.

(The ignored exceptions are still `MissingFile` and `NoWhereFromValue`,
but `UnsupportedFileSystemObject` should perhaps be ignored as well.)