Don’t ignore exceptions in base paths
If processing a path that has been explicitly specified as a base path
causes an exception that can usually be ignored, do not ignore it.

If a random file that was found somewhere in a directory tree the user
has specified doesn’t have a “where from” value or disappears before
the value can be read, that’s not something worth notifying the user
about, but if a path they explicitly specified on the command line
doesn’t exist or doesn’t have a “where from” value, the user will want
to know about that.