Make the application runnable
Add a `run()` function that runs the application, so that it can
actually be used from the command line, not just tested.

This has been added in a rush because there’s a job I’d like to apply
for as soon as possible, and I want to mention the application on my
résumé, so it should probably actually be able to run.